requirements.txt
Django==3.0.9
django-filter==2.3.0
djangorestframework==3.11.0
django-environ==0.4.5
importlib-metadata==1.7.0
Markdown==3.2.2
gunicorn==20.0.4
Pillow==7.2.0
psycopg2==2.8.5